Extract reusable `FeedStoreSpecs` helper methods into a shared scope so it can be used only by other `FeedStore` implementation tests
