+++ EssentialFeed/EssentialFeedTests/FeedCache/XCTestCase+FeedStoreSpecs.swift
@@ -1,0 +1,68 @@
+//
+//  XCTestCase+FeedStoreSpecs.swift
+//  EssentialFeedTests
+//
+//  Created by Radosław Serek on 17/11/2022.
+//
+
+import EssentialFeed
+import XCTest
+
+extension FeedStoreSpecs where Self: XCTestCase {
+    @discardableResult
+    func insert(feed: [LocalFeedImage], timestamp: Date, with sut: FeedStore) -> Error? {
+        let exp = XCTestExpectation(description: "Wait for retrieve completion")
+        var receivedError: Error?
+        sut.insert(feed, timestamp: timestamp) { insertionError in
+            receivedError = insertionError
+            exp.fulfill()
+        }
+        
+        wait(for: [exp], timeout: 1.0)
+        
+        return receivedError
+    }
+    
+    @discardableResult
+    func deleteCache(from sut: FeedStore) -> Error? {
+        let exp = XCTestExpectation(description: "Wait for delete completion")
+        
+        var deletionError: Error?
+        sut.deleteCachedFeed { error in
+            deletionError = error
+            exp.fulfill()
+        }
+        
+        wait(for: [exp], timeout: 1.0)
+
+        return deletionError
+    }
+    
+    func expect(_ sut: FeedStore, toRetrieve expectedResult: RetrieveCachedFeedResult, file: StaticString = #filePath, line: UInt = #line) {
+        let exp = XCTestExpectation(description: "Wait for cache retrieval")
+
+        sut.retrieve { retrievedResult in
+            switch (expectedResult, retrievedResult) {
+            case let (.found(expectedFeed, expectedTimestamp), .found(retrievedFeed, retrievedTimestamp)):
+                XCTAssertEqual(expectedFeed, retrievedFeed, file: file, line: line)
+                XCTAssertEqual(expectedTimestamp, retrievedTimestamp, file: file, line: line)
+                
+            case (.empty, .empty),
+                (.failure, .failure):
+                break
+                
+            default:
+                XCTFail("Expected to retrieve \(expectedResult), got  \(retrievedResult) instead", file: file, line: line)
+            }
+            exp.fulfill()
+        }
+
+        wait(for: [exp], timeout: 1.0)
+    }
+    
+    func expect(_ sut: FeedStore, toRetrieveTwice expectedResult: RetrieveCachedFeedResult, file: StaticString = #filePath, line: UInt = #line) {
+        expect(sut, toRetrieve: expectedResult, file: file, line: line)
+        expect(sut, toRetrieve: expectedResult, file: file, line: line)
+    }
+
+}
